+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.8.10
 + Added support for the Adult Filter in the Mob Slaughter Factory
++ The Enchantment Extractor will keep items in the input slot until they are fully disenchanted
 
 # 1.8.9
 + Fixed Server Side Crash with the Mob Crusher