+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Added support for the Adult Filter in the Mob Slaughter Factory
 + The Enchantment Extractor will keep items in the input slot until they are fully disenchanted
 + The Enchantment Extractor will now accept enchanted books with multiple enchantments and split them up
++ Fluids without buckets are now properly displayed in the Black Hole Unit
 
 # 1.8.9
 + Fixed Server Side Crash with the Mob Crusher